requirements.txt
robotframework
robotframework-seleniumlibrary==6.1.0
robotframework-faker==5.0.0
xlrd
selenium==4.9.1
robotframework-robocop==4.2.1
robotframework-pabot==2.16.0
robotframework-datadriver
robotframework-datadriver[XLS]
robotframework-retryfailed==0.2.0
robotframework-pdf2textlibrary
robotframework-screencaplibrary
requests
pyautogui
allure-robotframework
pywin32
openpyxl
